nersc_pvweb
a web based launcher for ParaView Web jobs on NERSC. The main
component is a javascript library that creates a UI for submitting
jobs, managing, and connecting to PV web jobs running on HPC
systems. There is also a session manager and a small collection of
server side scripts that handle the details of launching pvserver
and a twisted client through which a browser can connect. Apache
configuration examples are also included.